добавил всплывающие окна для авторизации
